requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
psycopg2-binary==2.9.9
django-cors-headers
djangorestframework
